Update charts, dashboard and also create sidebar with multiple pages
